+++ test/hierarch/bigint_primary_key_test.exs
@@ -4,9 +4,9 @@ defmodule BigintPrimaryKeyTest do
   setup_all do
     organizations = create_organizations()
 
-    on_exit fn ->
+    on_exit(fn ->
       Repo.delete_all(Organization)
-    end
+    end)
 
     {:ok, organizations}
   end
@@ -15,9 +15,11 @@ defmodule BigintPrimaryKeyTest do
     b = Map.get(organizations, "A.B")
     c = Map.get(organizations, "A.B.C")
 
-    descendants = c
-                  |> Organization.parent
-                  |> Repo.one
+    descendants =
+      c
+      |> Organization.parent()
+      |> Repo.one()
+
     assert descendants == b
   end
 
@@ -25,9 +27,11 @@ defmodule BigintPrimaryKeyTest do
     b = Map.get(organizations, "A.B")
     d = Map.get(organizations, "A.D")
 
-    siblings = b
-               |> Organization.siblings
-               |> Repo.all
+    siblings =
+      b
+      |> Organization.siblings()
+      |> Repo.all()
+
     assert siblings == [d]
   end
 
@@ -37,9 +41,11 @@ defmodule BigintPrimaryKeyTest do
     c = Map.get(organizations, "A.B.C")
     d = Map.get(organizations, "A.D")
 
-    descendants = a
-                  |> Organization.descendants
-                  |> Repo.all
+    descendants =
+      a
+      |> Organization.descendants()
+      |> Repo.all()
+
     assert descendants == [b, c, d]
   end
 end