+++ test/hierarch/root_test.exs
@@ -5,9 +5,9 @@ defmodule Hierarch.RootTest do
     Repo.delete_all(Catelog)
     catelogs = create_catelogs()
 
-    on_exit fn ->
+    on_exit(fn ->
       Repo.delete_all(Catelog)
-    end
+    end)
 
     {:ok, catelogs}
   end
@@ -15,20 +15,24 @@ defmodule Hierarch.RootTest do
   describe "root/1" do
     test "returns its root", catelogs do
       science = Map.get(catelogs, "Top.Science")
-      top     = Map.get(catelogs, "Top")
+      top = Map.get(catelogs, "Top")
 
-      root = science
-             |> Catelog.root
-             |> Repo.one
+      root =
+        science
+        |> Catelog.root()
+        |> Repo.one()
+
       assert root == top
     end
 
     test "returns itself if it is the root", catelogs do
       top = Map.get(catelogs, "Top")
 
-      root = top
-             |> Catelog.root
-             |> Repo.one
+      root =
+        top
+        |> Catelog.root()
+        |> Repo.one()
+
       assert root == top
     end
   end