+++ test/support/repo.ex
@@ -1,4 +1,6 @@
 defmodule Dummy.Repo do
   @moduledoc false
-  use Ecto.Repo, otp_app: :hierarch
+  use Ecto.Repo,
+    otp_app: :hierarch,
+    adapter: Ecto.Adapters.Postgres
 end