+++ test/test_helper.exs
@@ -4,6 +4,7 @@ defmodule Hierarch.TestCase do
   using(opts) do
     quote do
       use ExUnit.Case, unquote(opts)
+      import unquote(__MODULE__), only: [assert_match: 2]
       alias Dummy.{Repo, Catelog, Organization}
 
       def create_catelogs do
@@ -23,12 +24,17 @@ defmodule Hierarch.TestCase do
           "Top.Collections.Pictures.Astronomy.Astronauts"
         ]
 
-        Enum.reduce(catelogs_list, %{}, fn (name, acc) ->
+        Enum.reduce(catelogs_list, %{}, fn name, acc ->
           parent_name = Hierarch.LTree.parent_path(name)
           parent = Map.get(acc, parent_name)
 
-          catelog = Catelog.build_child_of(parent, %{name: name}) |> Repo.insert!
-          Map.put acc, name, catelog
+          catelog =
+            case parent do
+              nil -> Catelog.build(%{name: name}) |> Repo.insert!()
+              _ -> Catelog.build_child_of(parent, %{name: name}) |> Repo.insert!()
+            end
+
+          Map.put(acc, name, catelog)
         end)
       end
 
@@ -37,15 +43,20 @@ defmodule Hierarch.TestCase do
           "A",
           "A.B",
           "A.B.C",
-          "A.D",
+          "A.D"
         ]
 
-        Enum.reduce(organizations_list, %{}, fn (name, acc) ->
+        Enum.reduce(organizations_list, %{}, fn name, acc ->
           parent_name = Hierarch.LTree.parent_path(name)
           parent = Map.get(acc, parent_name)
 
-          organization = Organization.build_child_of(parent, %{name: name}) |> Repo.insert!
-          Map.put acc, name, organization
+          organization =
+            case parent do
+              nil -> Organization.build(%{name: name}) |> Repo.insert!()
+              _ -> Organization.build_child_of(parent, %{name: name}) |> Repo.insert!()
+            end
+
+          Map.put(acc, name, organization)
         end)
       end
     end
@@ -55,15 +66,32 @@ defmodule Hierarch.TestCase do
     Ecto.Adapters.SQL.Sandbox.mode(Dummy.Repo, {:shared, self()})
     :ok = Ecto.Adapters.SQL.Sandbox.checkout(Dummy.Repo)
   end
+
+  @doc """
+  ## Example
+
+  ```
+  assert_match [1, 2], [2, 1]
+  ```
+  """
+  @spec assert_match(list(), list()) :: term()
+  defmacro assert_match(list, another_list) do
+    quote location: :keep do
+      set = MapSet.new(unquote(list))
+      another_list = MapSet.new(unquote(another_list))
+
+      assert set == another_list
+    end
+  end
 end
 
 if System.get_env("CI") == "true" do
-  junit_folder = Mix.Project.build_path() <> "/junit/#{Mix.Project.config[:app]}"
+  junit_folder = Mix.Project.build_path() <> "/junit/#{Mix.Project.config()[:app]}"
   File.mkdir_p!(junit_folder)
   :ok = Application.put_env(:junit_formatter, :report_dir, junit_folder)
 
-  ExUnit.configure formatters: [JUnitFormatter, ExUnit.CLIFormatter]
+  ExUnit.configure(formatters: [JUnitFormatter, ExUnit.CLIFormatter])
 end
 
-{:ok, _} = Dummy.Repo.start_link
+{:ok, _} = Dummy.Repo.start_link()
 ExUnit.start()