+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## v0.2.0
+- rewrite
+- qurey descendants, ancestors, children of a query, resolve: #4
+- infer __hierarch__ info from schema, resolve: #3
+
 ## v0.1.2
 - path_column can be `NULL`